Number guessing game
in this file I wirtten a program for Number Guessing Game,The game has a clean and user-friendly interface. Its allow users to input their name and start the game. It has generate a random number between 1 and 100. It is allow  to users  guess the number within 10 attempts. It should provide feedback to the user after each guess, letting them know if their guess was too high or too low. It should declare the user as the winner if they guess the number correctly. It should give the user the option to play again or quit the game.